Redox: Use c_long instead of usize for off_t

+++ src/redox.rs
@@ -4,7 +4,7 @@ pub type c_ulong = u64;
 
 pub type wchar_t = i16;
 
-pub type off_t = usize;
+pub type off_t = c_long;
 pub type mode_t = u16;
 pub type time_t = i64;
 pub type pid_t = usize;
